delete unnecessary pipe, moved type to separate file
