+++ src/app/pages/movie-details/movie-details.component.html
@@ -20,7 +20,10 @@
           <mdb-detailed-page-hero-rating
             class="detailed-hero--rating"
             *ngIf="details.rating"
-            [ratingPostConfig]="details | toRatingPostConfig"
+            [ratingPostConfig]="{
+              type: details.mediaType,
+              id: details.id
+            }"
             [value]="details.rating"></mdb-detailed-page-hero-rating>
           <mdb-detailed-page-hero-section
             *ngIf="details.releaseDate"
