+++ src/app/pages/movie-details/movie-details.module.ts
@@ -4,16 +4,9 @@ import { MovieDetailsComponent } from './movie-details.component';
 import { PreloaderModule } from '../../components/preloader/preloader.module';
 import { DetailedPageModule } from '../../components/detailed-page/detailed-page.module';
 import { FilterModule } from '../../pipes/filter/filter.module';
-import { ToRatingPostConfigModule } from 'src/app/pipes/to-rating-post-config/to-rating-post-config.module';
 
 @NgModule({
   declarations: [MovieDetailsComponent],
-  imports: [
-    CommonModule,
-    PreloaderModule,
-    DetailedPageModule,
-    FilterModule,
-    ToRatingPostConfigModule,
-  ],
+  imports: [CommonModule, PreloaderModule, DetailedPageModule, FilterModule],
 })
 export class MovieDetailsModule {}
