--- src/app/pipes/to-rating-post-config/to-rating-post-config.module.ts
@@ -1,10 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { ToRatingPostConfigPipe } from './to-rating-post-config.pipe';
-
-@NgModule({
-  declarations: [ToRatingPostConfigPipe],
-  exports: [ToRatingPostConfigPipe],
-  imports: [CommonModule],
-})
-export class ToRatingPostConfigModule {}
