+++ src/app/types/rating-post-config.ts
@@ -1,0 +1,6 @@
+import { CanBeRatedMediaType } from './media-type';
+
+export interface RatingPostConfig {
+  type: CanBeRatedMediaType;
+  id: number;
+}
